Updates Email of all active sysadmins post-refresh
When a sandbox is created or refreshed, specifying `XC_SandboxPostCopy` as the Apex class to execute when the process is complete will update the Email address of all of the Users with the System Adminsitrator Profile, such that the email addresses are correct.

Example:  developer.person=xcentium.com@example.com will be updated to developer.person@xcentium.com.

This allows system admins to immediately begin working in the new sandbox, without having to request the intervention of the admin who initiated the creation/refresh of the sandbox.